requirements.txt
absl-py==2.4.0
astunparse==1.6.3
blinker==1.9.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
cryptography==47.0.0
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
gast==0.7.0
google-pasta==0.2.0
grpcio==1.80.0
h5py==3.16.0
idna==3.13
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
keras==3.14.0
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.1.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.19.0
packaging==26.2
pandas==3.0.2
passlib==1.7.4
protobuf==4.25.9
pyarrow==24.0.0
pycparser==3.0
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.1
rich==15.0.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.3.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
Werkzeug==3.1.8
wrapt==2.1.2
xgboost==3.2.0
